+++ src/main/java/org/morphix/convert/ConversionEngine.java
@@ -10,21 +10,23 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package org.morphix.convert.strategy;
+package org.morphix.convert;
+
+import org.morphix.convert.context.ConversionContext;
 
 /**
- * Deep name path strategy.
- * <p>
- * TODO: implement this
+ * Interface for a conversion engine that can convert values from one type to another.
  *
  * @author Radu Sebastian LAZIN
  */
-public class DeepNamePathStrategy extends NamePathStrategy {
+public interface ConversionEngine {
 
 	/**
-	 * Default constructor.
+	 * Converts the given value to a different type based on the provided conversion context.
+	 *
+	 * @param value the value to be converted
+	 * @param ctx the conversion context containing information about the conversion
+	 * @return the converted value
 	 */
-	public DeepNamePathStrategy() {
-		// empty
-	}
+	Object convert(Object value, ConversionContext ctx);
 }