+++ src/main/java/org/morphix/convert/ObjectConverter.java
@@ -24,7 +24,7 @@ import org.morphix.convert.annotation.Src;
 import org.morphix.convert.function.ConvertFunction;
 import org.morphix.convert.function.InstanceConvertFunction;
 import org.morphix.convert.function.SimpleConverter;
-import org.morphix.convert.strategy.ConversionStrategy;
+import org.morphix.convert.strategy.FieldFinderStrategy;
 import org.morphix.lang.function.InstanceFunction;
 import org.morphix.reflection.ExtendedField;
 
@@ -83,7 +83,7 @@ public class ObjectConverter<S, D> implements
 	 *
 	 * @return a list of strategies with which to find the fields in the source
 	 */
-	public List<ConversionStrategy> getStrategies() {
+	public List<FieldFinderStrategy> getStrategies() {
 		return this.configuration.getStrategies();
 	}
 
@@ -165,18 +165,18 @@ public class ObjectConverter<S, D> implements
 	 * @param destination destination object
 	 */
 	private void mainConvert(final S source, final D destination) {
-		List<ExtendedField> dstFields = ConversionStrategy.findFields(destination);
+		List<ExtendedField> dstFields = FieldFinderStrategy.findFields(destination);
 		if (dstFields.isEmpty()) {
 			return;
 		}
-		List<ExtendedField> srcFields = ConversionStrategy.findFields(source);
+		List<ExtendedField> srcFields = FieldFinderStrategy.findFields(source);
 		if (srcFields.isEmpty()) {
 			return;
 		}
 		for (ExtendedField dfo : dstFields) {
 			String srcFieldName = getSourceFieldName(dfo, source);
 			// apply source field finding strategies
-			for (ConversionStrategy strategy : getStrategies()) {
+			for (FieldFinderStrategy strategy : getStrategies()) {
 				ExtendedField sfo = strategy.find(source, srcFields, srcFieldName);
 				if (sfo.hasObject()) {
 					convertField(sfo, dfo);