+++ src/test/java/org/morphix/convert/strategy/FieldFinderStrategyTest.java
@@ -22,11 +22,11 @@ import org.junit.jupiter.api.Test;
 import org.morphix.reflection.ExtendedField;
 
 /**
- * Test class for {@link ConversionStrategy}.
+ * Test class for {@link FieldFinderStrategy}.
  *
  * @author Radu Sebastian LAZIN
  */
-class ConversionStrategyTest {
+class FieldFinderStrategyTest {
 
 	public static class A {
 		String x;
@@ -48,7 +48,7 @@ class ConversionStrategyTest {
 		// empty
 	}
 
-	public static class TestStrategy implements ConversionStrategy {
+	public static class TestStrategy implements FieldFinderStrategy {
 		@Override
 		public <T> ExtendedField find(final T source, final List<ExtendedField> fields, final String sourceFieldName) {
 			return null;